Feat/uni v4 update

+++ src/ldf/DoubleGeometricDistribution.sol
@@ -74,7 +74,7 @@ contract DoubleGeometricDistribution is ILiquidityDensityFunction {
         external
         pure
         override
-        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint128 swapLiquidity)
+        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint256 swapLiquidity)
     {
         (
             int24 minTick,
