+++ test/mocks/MockLDF.sol
@@ -64,7 +64,7 @@ contract MockLDF is ILiquidityDensityFunction {
         external
         view
         override
-        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint128 swapLiquidity)
+        returns (bool success, int24 roundedTick, uint256 cumulativeAmount, uint256 swapLiquidity)
     {
         (int24 minTick, int24 length, uint256 alphaX96, ShiftMode shiftMode) =
             LibGeometricDistribution.decodeParams(twapTick, key.tickSpacing, useTwap, ldfParams);
